PHOENIX-7022 :- Add new columns TTL and ROW_KEY_PREFIX (#1660)
* PHOENIX-7022 :- Add TTL and ROWKEY_PREFIX Column and Use TTL instead of PHOENIX_TTL for storing TTL in SYSCAT

+++ phoenix-core/src/main/java/org/apache/phoenix/coprocessor/CompactionScanner.java
@@ -86,7 +86,7 @@ public class CompactionScanner implements InternalScanner {
             long maxLookbackInMillis,
             byte[] emptyCF,
             byte[] emptyCQ,
-            long phoenixTTL,
+            int phoenixTTL,
             boolean isSystemTable) {
         this.storeScanner = storeScanner;
         this.region = env.getRegion();
